Allow configuration of the image via build variables

+++ packer/scripts/wisebox.sh
@@ -1,16 +1,6 @@
 #!/bin/bash
 set -e
 
-WB_USERNAME="wisebox"
-WB_PASSWORD="wisebox"
-WB_HOSTNAME="wisebox"
-WB_PRETTY_HOSTNAME="WISEBox"
-WB_LOCALE="en_GB.UTF-8"
-WB_TIMEZONE="Europe/London"
-WB_WIFI_COUNTRY="GB"
-WB_LOGGER_VERSION_TAG="v1.3.0"
-WB_LOGSERVER_VERSION_TAG="v1.1.0"
-
 apt-get install -y \
     bluez \
     bridge-utils \
@@ -31,6 +21,10 @@ pip3 install pipenv
 chown -R root:root /wisebox-files
 cp -fr --preserve=mode /wisebox-files/* /
 rm -rf /wisebox-files
+if [ -e "/etc/wisebox/wisebox.override.yml" ]; then
+  rm -f /etc/wisebox/wisebox.yml
+  mv /etc/wisebox/wisebox.override.yml /etc/wisebox/wisebox.yml
+fi
 
 usermod -l ${WB_USERNAME} pi
 usermod -m -d /home/${WB_USERNAME} ${WB_USERNAME}
@@ -44,15 +38,21 @@ raspi-config nonint do_change_locale ${WB_LOCALE}
 raspi-config nonint do_change_timezone ${WB_TIMEZONE}
 echo "PRETTY_HOSTNAME=${WB_PRETTY_HOSTNAME}" > /etc/machine-info
 
-sed -i -e "s/WB_WIFI_COUNTRY/${WB_WIFI_COUNTRY}/" /etc/init.d/wisebox-once.sh
+sed -i \
+    -e "s/WB_WIFI_COUNTRY/${WB_WIFI_COUNTRY}/" \
+    -e "s/WB_ENABLE_SSH_SERVER/${WB_ENABLE_SSH_SERVER}/" \
+    -e "s/WB_ENABLE_HW_CLOCK/${WB_ENABLE_HW_CLOCK}/" \
+    /etc/init.d/wisebox-once.sh
 systemctl enable wisebox-once.sh
 
-echo "dtoverlay=i2c-rtc,ds3231" >> /boot/config.txt
-patch /lib/udev/hwclock-set /lib/udev/hwclock-set.patch
+if [ "${WB_ENABLE_HW_CLOCK}" -eq 1 ]; then
+  echo "dtoverlay=i2c-rtc,ds3231" >> /boot/config.txt
+  patch /lib/udev/hwclock-set /lib/udev/hwclock-set.patch
+  apt-get -y remove fake-hwclock
+  update-rc.d -f fake-hwclock remove
+  systemctl disable fake-hwclock
+fi
 rm -f /lib/udev/hwclock-set.patch
-apt-get -y remove fake-hwclock
-update-rc.d -f fake-hwclock remove
-systemctl disable fake-hwclock
 
 WB_LOGGER_DIR=/tmp/logger
 git clone git://github.com/Horizon-WISEBox/wisebox-logger.git ${WB_LOGGER_DIR}
@@ -66,13 +66,15 @@ ln -s /etc/wisebox/wisebox-logger.service \
     /etc/systemd/system/wisebox-logger.service
 systemctl enable wisebox-logger.service
 
-patch /etc/bluetooth/main.conf /etc/bluetooth/main.conf.patch
+if [ "${WB_ENABLE_BLUETOOTH_NETWORKING}" -eq 1 ]; then
+  patch /etc/bluetooth/main.conf /etc/bluetooth/main.conf.patch
+  patch /etc/dhcp/dhcpd.conf /etc/dhcp/dhcpd.conf.patch
+  echo bnep >> /etc/modules
+  systemctl enable blueagent5.service
+  systemctl enable bt-pan.service
+fi
 rm -f /etc/bluetooth/main.conf.patch
-patch /etc/dhcp/dhcpd.conf /etc/dhcp/dhcpd.conf.patch
 rm -f /etc/dhcp/dhcpd.conf.patch
-echo bnep >> /etc/modules
-systemctl enable blueagent5.service
-systemctl enable bt-pan.service
 
 WB_LOGSERVER_DIR=/tmp/logfile-server
 patch /etc/lighttpd/conf-available/10-fastcgi.conf \
@@ -90,7 +92,3 @@ rm -rf /var/www/html/*
 mv static/* /var/www/html
 cd /
 rm -rf ${WB_LOGSERVER_DIR}
-
-# WISEBox Config
-#https://unix.stackexchange.com/questions/211544/run-top-instead-of-getty
-#https://raymii.org/s/tutorials/Run_software_on_tty1_console_instead_of_login_getty.html
